# packages/contracts/tsp-output/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  description: |-
    Project API - RESTful API following modern best practices

    Key Features:
    - URL versioning (/v1)
    - Multi-tenant architecture (path-based with /orgs/{orgId})
    - Comprehensive pagination with navigation links
    - Structured error responses with detailed validation
    - Soft delete support with restore capability
    - Batch operations for bulk processing
    - Field selection and relation inclusion
    - Full-text search and filtering
  version: 0.0.0
...
paths:
...
  /v1/orgs/{orgId}/roles:
    get:
      operationId: TenantRoles_list
      summary: List tenant roles
      description: |-
        List all roles in a tenant

        Returns both system roles and custom roles for this tenant.
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: Page number (1-indexed)
          schema:
            type: integer
            format: int32
            default: 1
          explode: false
        - name: pageSize
          in: query
          required: false
          description: Number of items per page
          schema:
            type: integer
            format: int32
            default: 50
          explode: false
        - name: isSystemRole
          in: query
          required: false
          description: Filter by system role status
          schema:
            type: boolean
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/RoleListResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Roles
    post:
      operationId: TenantRoles_create
      summary: Create tenant role
      description: |-
        Create a tenant-scoped role

        Creates a custom role for this specific tenant.
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
      tags:
        - Roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
        description: Role data
...
  /v1/roles:
    get:
      operationId: GlobalRoles_list
      summary: List global roles
      description: |-
        List all global roles

        Returns global roles (no tenant scope) for the default application.
      parameters:
        - name: page
          in: query
          required: false
          description: Page number (1-indexed)
          schema:
            type: integer
            format: int32
            default: 1
          explode: false
        - name: pageSize
          in: query
          required: false
          description: Number of items per page
          schema:
            type: integer
            format: int32
            default: 50
          explode: false
        - name: isSystemRole
          in: query
          required: false
          description: Filter by system role status
          schema:
            type: boolean
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/RoleListResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Roles
    post:
      operationId: GlobalRoles_create
      summary: Create global role
      description: |-
        Create a global role

        Creates a role that applies across all tenants.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
      tags:
        - Roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
        description: Role data
  /v1/roles/{roleId}:
    get:
      operationId: GlobalRoles_get
      summary: Get role
      description: Get a role by ID
      parameters:
        - name: roleId
          in: path
          required: true
          description: Role ID
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/RoleResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Roles
    patch:
      operationId: GlobalRoles_update
      summary: Update role
      description: |-
        Update a role

        System roles cannot be renamed but their permissions can be modified.
      parameters:
        - name: roleId
          in: path
          required: true
          description: Role ID
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/RoleResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
        description: Role update data
    delete:
      operationId: GlobalRoles_delete
      summary: Delete role
      description: |-
        Delete a role

        System roles cannot be deleted.
        Fails if role is assigned to any users.
      parameters:
        - name: roleId
          in: path
          required: true
          description: Role ID
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
      tags:
        - Roles
...
components:
  schemas:
...
    CreateRoleRequest:
      type: object
      required:
        - name
        - permissions
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Role name
        description:
          type: string
          maxLength: 500
          description: Role description
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionInput'
          minItems: 1
          description: Permissions to grant
      description: Request to create a custom role
...
    Error:
      type: object
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          description: Array of detailed error information (e.g., validation errors)
        requestId:
          type: string
          description: Request ID for tracking and debugging
        documentationUrl:
          type: string
          description: URL to documentation about this error (optional)
      description: Standard error object
    ErrorDetail:
      type: object
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          description: The field that caused the error (for validation errors)
        code:
          type: string
          description: Specific error code for this detail
        message:
          type: string
          description: Human-readable error message
        metadata:
          type: object
          additionalProperties: {}
          description: Additional metadata about the error
      description: Detailed error information for a specific field or validation
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          allOf:
            - $ref: '#/components/schemas/Error'
          description: The error object
      description: Standard error response envelope
...
    Pagination:
      type: object
      required:
        - page
        - pageSize
        - totalPages
        - totalCount
        - hasNext
        - hasPrevious
      properties:
        page:
          type: integer
          format: int32
          description: Current page number (1-indexed)
        pageSize:
          type: integer
          format: int32
          description: Number of items per page
        totalPages:
          type: integer
          format: int32
          description: Total number of pages
        totalCount:
          type: integer
          format: int32
          description: Total number of items across all pages
        hasNext:
          type: boolean
          description: Whether there is a next page
        hasPrevious:
          type: boolean
          description: Whether there is a previous page
        links:
          allOf:
            - $ref: '#/components/schemas/PaginationLinks'
          description: Navigation links for pagination
        requestedPageSize:
          type: integer
          format: int32
          description: Original requested page size (included if pageSize was capped)
      description: Page-based pagination metadata for collection responses
    PaginationLinks:
      type: object
      required:
        - first
        - previous
        - next
        - last
      properties:
        first:
          type: string
          nullable: true
          description: URL for the first page
        previous:
          type: string
          nullable: true
          description: URL for the previous page (null if on first page)
        next:
          type: string
          nullable: true
          description: URL for the next page (null if on last page)
        last:
          type: string
          nullable: true
          description: URL for the last page
      description: Pagination links for navigating through paginated results
    Permission:
      type: object
      required:
        - resource
        - action
        - effect
      properties:
        resource:
          type: string
          description: Resource identifier (e.g., "users", "invoices", "settings")
          example: documents
        action:
          anyOf:
            - $ref: '#/components/schemas/PermissionAction'
            - type: string
          description: Action to perform on the resource
          example: update
        effect:
          allOf:
            - $ref: '#/components/schemas/PermissionEffect'
          description: Whether to allow or deny this action
        condition:
          allOf:
            - $ref: '#/components/schemas/PermissionCondition'
          description: Optional condition for when this permission applies
      description: |-
        Permission definition

        Permissions define what actions can be performed on resources.
        They support allow/deny effects and optional conditions.
    PermissionAction:
      type: string
      enum:
        - read
        - create
        - update
        - delete
        - manage
        - '*'
      description: Permission actions
    PermissionCondition:
      type: string
      enum:
        - ''
        - owner
        - shared
      description: |-
        Permission condition - when the permission applies

        Conditions enable dynamic, context-aware access control.
    PermissionEffect:
      type: string
      enum:
        - allow
        - deny
      description: |-
        Permission effect - whether to allow or deny access

        When a user has multiple roles, permissions are combined:
        - If ANY role denies an action, it is denied (deny wins)
        - If no role denies and ANY role allows, it is allowed
        - If no rules match, access is denied (default deny)
    PermissionInput:
      type: object
      required:
        - resource
        - action
      properties:
        resource:
          type: string
          description: Resource identifier
        action:
          type: string
          description: Action to perform
        effect:
          allOf:
            - $ref: '#/components/schemas/PermissionEffect'
          description: Effect (defaults to "allow" if not specified)
        condition:
          allOf:
            - $ref: '#/components/schemas/PermissionCondition'
          description: Optional condition
      description: Permission input for creating/updating roles
    ResponseMeta:
      type: object
      required:
        - requestId
        - timestamp
      properties:
        requestId:
          type: string
          description: Unique request identifier for tracking and debugging
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the response was generated
        apiVersion:
          type: string
          description: API version used for this request
        tenantId:
          type: string
          description: Tenant/organization ID for multi-tenant requests
        tenantName:
          type: string
          description: Tenant/organization name for multi-tenant requests
        durationMs:
          type: integer
          format: int32
          description: Processing time in milliseconds
      description: Response metadata included in all API responses
    Role:
      type: object
      required:
        - id
        - applicationId
        - name
        - permissions
        - isSystemRole
        - isGlobalRole
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: 'Unique role identifier (format: role_abc123)'
          example: role_admin
        applicationId:
          type: string
          description: Application this role belongs to
          example: app_default
        tenantId:
          type: string
          description: Tenant this role belongs to (null for global roles)
          example: org_abc123
        name:
          type: string
          description: Role name
          example: Admin
        description:
          type: string
          description: Role description
          example: Full administrative access
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
          description: Permissions granted by this role
        isSystemRole:
          type: boolean
          description: Whether this is a system-defined role (cannot be deleted or renamed)
        isGlobalRole:
          type: boolean
          description: Whether this is a global role (tenantId is null)
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the role was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the role was last updated
        createdBy:
          type: string
          description: User who created this role
          example: usr_owner
      description: |-
        Role resource model

        Roles can be:
        - **Global roles**: App-scoped (tenantId is null), apply across all tenants
        - **Tenant roles**: Scoped to a specific organization

        Users can have multiple roles, and permissions are combined with
        deny-override semantics.
    RoleListResponse:
      type: object
      required:
        - data
        - pagination
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      description: Role list response
    RoleResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/Role'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      description: Single role response
...
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Role name
        description:
          type: string
          maxLength: 500
          description: Role description
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionInput'
          description: Permissions to grant (replaces existing permissions)
      description: Request to update a role
...
servers:
  - url: https://api.example.com
    description: Production server
    variables: {}
  - url: http://localhost:3001
    description: Development server
    variables: {}